backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
torch==2.11.0
torchvision==0.26.0
timm==0.9.16
facenet-pytorch==2.5.3
Pillow==10.3.0
numpy==1.26.4
python-dotenv==1.0.1